refactor: clean up test file
